Removed unnecessary code (we had two sets of genre methods), moved methods to their respective handler folders (from apiHandler)
